+++ src/overlays/actors/ovl_En_Okarina_Effect/z_en_okarina_effect.cpp
@@ -87,7 +87,7 @@ void EnOkarinaEffect_ManageStorm(EnOkarinaEffect* pthis, GlobalContext* globalCt
 		{
 			pthis->timer--;
 		}
-		osSyncPrintf("\nthis->timer=[%d]", pthis->timer);
+		osSyncPrintf("\nthis->timer=[%d]", (int)pthis->timer);
 		if(pthis->timer == 308)
 		{
 			osSyncPrintf("\n\n\n豆よ のびろ 指定\n\n\n"); // "Let's grow some beans"
